Refactor components to ui-ill-components
A number of components need to be shared by ui-ill-ra &
ui-plugin-connector-bldss, and will also need to be usable by future
connectors. This commit therefore refactors them, and their
translations, into a new ui-ill-components package.

We also add ui-ill-components as a dependency.

This commit also fixes a couple of small translation string errors.